fix(detector+demo): Browser-Performance — Rausch-Floor von Long auf Double + Overlay sofort beim Bildwechsel löschen
Root-Cause-Analyse (systematic-debugging) der „unglaublich langsamen" Demo:

1. detectPanels brauchte ~1166 ms/Seite (1084×1500) im Browser. Ursache: der in
   0.2.1 eingeführte localNoiseFloor nutzte LongArray + Long-Arithmetik pro Pixel —
   Kotlin/JS EMULIERT Long als Objekt, eine ~1.6-MP-Schleife kostet so ~1 s.
   Fix: Integral-Bilder als DoubleArray (JS-nativ, exakt bis 2^53 ≫ sumSq ≈ 1e11).
   → 1166 ms → 364 ms (−69 %), gemessen mit dem echten Bundle in node.
   Detektion bitidentisch (JVM-Golden-Benchmark unverändert 0.677/0.546/43).

2. Alte Panel-Boxen blieben bis zur fertigen (synchronen) Erkennung des NÄCHSTEN
   Bildes stehen → wirkte stark verzögert. Ursache: kein clearOverlay beim Wechsel.
   Fix in main.ts: Overlay + Count sofort beim loadSrc leeren; Erkennung erst nach
   einem Paint (doppeltes rAF), damit das neue Bild sofort erscheint statt erst
   nach der Rechenzeit.

Version 0.2.1 → 0.2.2.

+++ src/commonMain/kotlin/com/panela/comiccutter/GutterProfileCut.kt
@@ -148,15 +148,18 @@ object GutterProfileCut {
         val r = win / 2
         if (w <= win || h <= win) return 0.0
         val sw = w + 1
-        val sum = LongArray(sw * (h + 1))
-        val sumSq = LongArray(sw * (h + 1))
+        // Double (nicht Long) für die Integral-Bilder: Kotlin/JS emuliert Long als Objekt — eine
+        // Pro-Pixel-Long-Schleife über ~1.6 MP kostet im Browser ~1 s. Double ist JS-nativ und stellt
+        // die Summen exakt dar (sumSq ≤ 255²·Pixel ≈ 1e11 ≪ 2^53). Auf der JVM identisch.
+        val sum = DoubleArray(sw * (h + 1))
+        val sumSq = DoubleArray(sw * (h + 1))
         for (y in 0 until h) {
-            var rowSum = 0L
-            var rowSqr = 0L
+            var rowSum = 0.0
+            var rowSqr = 0.0
             val above = y * sw
             val cur = (y + 1) * sw
             for (x in 0 until w) {
-                val v = lum[y * w + x].toLong()
+                val v = lum[y * w + x].toDouble()
                 rowSum += v; rowSqr += v * v
                 sum[cur + x + 1] = sum[above + x + 1] + rowSum
                 sumSq[cur + x + 1] = sumSq[above + x + 1] + rowSqr
